efs_parsing/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
autoflake==2.3.1
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.8
cryptography==43.0.3
dataclasses-json==0.6.7
deepsearch-glm==1.0.0
dill==0.3.9
distro==1.9.0
docling==2.12.0
docling-core==2.11.0
docling-ibm-models==3.1.0
docling-parse==3.0.0
easyocr==1.7.2
emoji==2.14.0
et_xmlfile==2.0.0
eval_type_backport==0.2.0
fastapi==0.115.3
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
imageio==2.36.1
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.1.143
lazy_loader==0.4
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.55.0
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==10.4.0
pip==24.2
ppt2pdf==0.0.3
prompt-toolkit==1.0.14
propcache==0.2.0
psutil==6.1.0
pybase64==1.4.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.6.1
pyfiglet==1.0.2
pyflakes==3.2.0
Pygments==2.18.0
PyInquirer==1.0.3
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.0
python-bidi==0.6.3
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.2
pywin32==307
PyYAML==6.0.2
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
Rtree==1.3.0
safetensors==0.4.5
scikit-image==0.25.0
scipy==1.14.1
semchunk==2.2.0
setuptools==75.6.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.41.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tifffile==2024.12.12
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
transformers==4.47.0
typer==0.12.5
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
unstructured==0.16.5
unstructured-client==0.27.0
urllib3==2.2.3
uvicorn==0.32.0
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.17.2
